Untested refactoring, mainly of camera code
