+++ Plugin.cs
@@ -101,16 +101,6 @@ public class VSVRMod : BaseUnityPlugin
 
     public void InitialSessionSetup()
     {
-        if (vrCameraManager != null)
-        {
-            controllerHeadset.OnWorn -= vrCameraManager.SetupUI;
-            controllerHeadset.OnRemoved -= vrCameraManager.RevertUI;
-        }
-        if (uiContainer != null)
-        {
-            vrGestureRecognizer.Nodded -= uiContainer.basicUIManager.headMovementTracker.Nod;
-            vrGestureRecognizer.HeadShaken -= uiContainer.basicUIManager.headMovementTracker.Headshake;
-        }
         logger.LogInfo("Starting session setup");
         vrCameraManager = new(sessionScene);
         logger.LogInfo("Session setup: created camera manager");
@@ -123,14 +113,6 @@ public class VSVRMod : BaseUnityPlugin
         vrGestureRecognizer.HeadShaken += uiContainer.basicUIManager.headMovementTracker.Headshake;
         logger.LogInfo("Session setup: setup gestures");
 
-        controllerHeadset.OnWorn += vrCameraManager.SetupUI;
-        controllerHeadset.OnRemoved += vrCameraManager.RevertUI;
-        logger.LogInfo("Session setup: OnWorn and OnRemoved");
-        if (!VRConfig.taskGradient.Value)
-        {
-            vrCameraManager.DisableTaskGradient();
-            logger.LogInfo("Session setup: Disabled task gradients");
-        }
         inSession = true;
     }
 
@@ -151,8 +133,7 @@ public class VSVRMod : BaseUnityPlugin
                 controllerHeadset.Update();
             }
             uiContainer.Interact();
-            vrCameraManager.CameraControls();
-            vrCameraManager.CenterCameraIfFar();
+            vrCameraManager.Update();
         }
         else
         {
