+++ UI/SpecifcUI/ScoreboardUI.cs
@@ -1,7 +1,7 @@
 ﻿using UnityEngine.SceneManagement;
 using UnityEngine;
 
-namespace VSVRMod2.UI;
+namespace VSVRMod2.UI.SpecifcUI;
 public class ScoreboardUIManager : UIManager
 {
     private struct Scoreboard
